+++ src/components/nvi/dialogs/role-contact-information/CuratorName.tsx
@@ -1,13 +1,13 @@
 import { Link as MuiLink, Typography } from '@mui/material';
 import { Link } from 'react-router';
-import { getResearchProfilePath } from '../../../utils/urlPaths';
+import { getResearchProfilePath } from '../../../../utils/urlPaths';
 
-interface CuratorContactInformationProps {
+interface CuratorNameProps {
   name: string;
   cristinId?: string;
 }
 
-export const CuratorName = ({ name, cristinId }: CuratorContactInformationProps) => {
+export const CuratorName = ({ name, cristinId }: CuratorNameProps) => {
   if (cristinId) {
     return (
       <MuiLink component={Link} to={getResearchProfilePath(cristinId)}>
