+++ src/components/nvi/table/rows/NviAdminReportingStatusRow.tsx
@@ -1,7 +1,7 @@
 import { TableCell, TableRow } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { InstitutionReport } from '../../../../types/nvi.types';
-import { ViewContactInfoButton } from '../../../_molecules/buttons/ViewContactInformationButton';
+import { ViewContactInfoButton } from '../../../_molecules/buttons/ViewContactInfoButton';
 import { PercentageWithIcon } from '../../../_molecules/PercentageWithIcon';
 import { HorizontalBox } from '../../../styled/Wrappers';
 import { CenteredTableCell } from '../../../tables/table-styles';
